Fixes temp files not being deleted issue
As reported in MULE-8934 The file was just being flushed and remained open when the file was due to be deleted, it would more often not fail as the if was still in use. This can cause huge startup times as each message gets checked for completion each time the mule service starts. 

Changed the serialize method to close the file - now the temp files car now be deleted.